requirements.txt

# Core Dependencies
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# Financial Data Processing
ta-lib==0.4.25
yfinance==0.2.18

# Snowflake Integration
snowflake-connector-python==3.3.1

# Visualization
plotly==5.16.1

# Data Science & Analysis
scikit-learn==1.3.0
scipy==1.11.1

# Optional: For enhanced features
websockets==11.0.3
apscheduler==3.10.4
